remove old test_secrets, mv tests to test_operations
